# docs/superpowers/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Hunter Platform API",
    "version": "1.4.1",
    "description": "猎头中介 API 平台 — HTTP-only API for external AI agents. Detailed docs in /v1/skill.md."
  },
  "servers": [
    {
      "url": "https://api.hunter-platform.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local dev"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key issued by POST /v1/auth/register. Pass as `Authorization: Bearer hp_live_...`."
      }
    },
    "schemas": {
      "Ok": {
        "type": "object",
        "required": [
          "ok",
          "data"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "data": {}
        }
      },
      "Err": {
        "type": "object",
        "required": [
          "ok",
          "error"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "UNAUTHORIZED",
                  "FORBIDDEN",
                  "NOT_FOUND",
                  "INVALID_PARAMS",
                  "INVALID_STATE",
                  "DUPLICATE_REQUEST",
                  "CONTACT_TAKEN",
                  "INSUFFICIENT_QUOTA",
                  "RATE_LIMITED",
                  "INVALID_CHARSET",
                  "INVALID_CONTENT_TYPE",
                  "INVALID_JSON",
                  "PAYLOAD_TOO_LARGE",
                  "NOT_IMPLEMENTED",
                  "INTERNAL_ERROR"
                ]
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      },
      "UserType": {
        "type": "string",
        "enum": [
          "candidate",
          "headhunter",
          "employer"
        ]
      },
      "UserStatus": {
        "type": "string",
        "enum": [
          "active",
          "suspended",
          "deleted"
        ]
      },
      "UserPublic": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "user_type": {
            "$ref": "#/components/schemas/UserType"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "quota_per_day": {
            "type": "integer"
          },
          "quota_used": {
            "type": "integer"
          },
          "quota_reset_at": {
            "type": "string",
            "format": "date-time"
          },
          "reputation": {
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/UserStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ActionHistoryEntry": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "user_id": {
            "type": "string"
          },
          "action_type": {
            "type": "string"
          },
          "target_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "target_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "request_summary_json": {
            "type": [
              "string",
              "null"
            ]
          },
          "response_summary_json": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "success",
              "error"
            ]
          },
          "error_code": {
            "type": [
              "string",
              "null"
            ]
          },
          "duration_ms": {
            "type": [
              "integer",
              "null"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid API key",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Err"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden for this user type / scope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Err"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Err"
            }
          }
        }
      },
      "InvalidParams": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Err"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "ApiKey": []
    }
  ],
  "paths": {
...
    "/v1/users/{id}/status": {
      "get": {
        "summary": "Get user status (quota / reputation / type)",
        "security": [
          {
            "ApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Ok"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserPublic"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/v1/users/{id}/history": {
      "get": {
        "summary": "Get caller's own action history",
        "security": [
          {
            "ApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            }
          },
          {
            "name": "since",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Only return entries with created_at >= this ISO 8601 timestamp"
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Ok"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/ActionHistoryEntry"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidParams"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  }
}